agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "maa-cli-agent"
version = "0.1.0"
description = "Polling agent that bridges maa-cli with the remote control server."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0,<0.28.0",
    "pydantic>=2.7.1,<2.8.0",
    "pyyaml>=6.0.1,<7.0.0",
    "typer>=0.12.3,<0.13.0"
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

